third_party/blink/web_tests/fast/canvas/canvas-filter-origin-clean.html

<svg style="display: block; width: 0; height: 0">
  <defs>
    <filter id="drop-shadow">
      <feOffset dx="0" dy="10" result="offset" in="SourceAlpha"></feOffset>
      <feFlood flood-color="currentcolor"></feFlood>
      <feComposite in2="offset" operator="in"></feComposite>
      <feMerge>
        <feMergeNode></feMergeNode>
        <feMergeNode in="SourceGraphic"></feMergeNode>
      </feMerge>
    </filter>
    <filter id="image">
      <feImage xlink:href="#circle" />
      <feColorMatrix type="matrix" values="0 0 0 1 0  0 0 0 1 0  0 0 0 1 0  0 0 0 0 1" />
    </filter>
    <filter id="merge-clean">
      <feColorMatrix type="matrix" values="0 0 0 1 0  0 0 0 1 0  0 0 0 1 0  0 0 0 0 1" />
      <feMerge>
        <feMergeNode></feMergeNode>
        <feMergeNode in="SourceGraphic"></feMergeNode>
        <feMergeNode in="SourceAlpha"></feMergeNode>
        <feMergeNode in="FillPaint"></feMergeNode>
        <feMergeNode in="StrokePaint"></feMergeNode>
      </feMerge>
    </filter>
    <circle id="circle" r="100" fill="blue" />
  </defs>
</svg>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script>

function assert_not_tainted(performCommands, description) {
  let ctx = document.createElement('canvas').getContext('2d');
  performCommands(ctx);
  assert_not_equals(ctx.getImageData(0, 0, 1, 1), null, description);
}

function assert_tainted(performCommands, description) {
  let ctx = document.createElement('canvas').getContext('2d');
  performCommands(ctx);
  assert_throws("SecurityError", () => ctx.getImageData(0, 0, 1, 1), description);
}

// CSS shorthand filters never taint the canvas.
test(function() {
  assert_not_tainted(ctx => {
    ctx.fillStyle = '#0f0';
    ctx.filter = 'brightness(0.5)';
    ctx.fillRect(5, 5, 10, 10);
  }, 'brightness(0.5)');

  assert_not_tainted(ctx => {
    ctx.filter = 'blur(5px)';
    ctx.fillRect(5, 5, 10, 10);
  }, 'blur(5px)');

  assert_not_tainted(ctx => {
    ctx.filter = 'hue-rotate(45deg) drop-shadow(16px 16px 20px blue)';
    ctx.fillRect(5, 5, 10, 10);
  }, 'hue-rotate(45deg) drop-shadow(16px 16px 20px blue)');
}, "CSS shorthand filters never taint the canvas");

// SVG reference filters taint the canvas according to
// https://drafts.fxtf.org/filter-effects/#tainted-filter-primitives

test(function() {
  assert_not_tainted(ctx => {
    ctx.filter = 'url(#merge-clean)';
    ctx.fillRect(5, 5, 10, 10);
  }, 'url(#merge-clean)');

  assert_not_tainted(ctx => {
    ctx.filter = 'blur(5px) url(#merge-clean) blur(5px)';
    ctx.fillRect(5, 5, 10, 10);
  }, 'blur(5px) url(#merge-clean) blur(5px)');
}, "Whitelisted SVG filters don't taint the canvas");

test(function() {
  assert_tainted(ctx => {
    ctx.filter = 'url(#drop-shadow)';
    ctx.fillRect(5, 5, 10, 10);
  }, 'url(#drop-shadow)');

  assert_tainted(ctx => {
    ctx.fillStyle = '#0f0';
    ctx.filter = 'url(#drop-shadow) brightness(0.5) url(#merge-clean)';
    ctx.fillRect(5, 5, 10, 10);
  }, 'url(#drop-shadow) brightness(0.5) url(#merge-clean)');

  assert_tainted(ctx => {
    ctx.fillStyle = '#0f0';
    ctx.filter = 'brightness(0.5) url(#drop-shadow)';
    ctx.fillRect(5, 5, 10, 10);
  }, 'brightness(0.5) url(#drop-shadow)');

  assert_tainted(ctx => {
    ctx.fillStyle = '#0f0';
    ctx.filter = 'brightness(0.5) url(#drop-shadow) brightness(0.5)';
    ctx.fillRect(5, 5, 10, 10);
  }, 'brightness(0.5) url(#drop-shadow) brightness(0.5)');

  assert_tainted(ctx => {
    ctx.fillStyle = '#0f0';
    ctx.filter = 'brightness(0.5) url(#drop-shadow) brightness(0.5)';
    ctx.fillRect(5, 5, 10, 10);
  }, 'brightness(0.5) url(#drop-shadow) brightness(0.5)');

  assert_tainted(ctx => {
    ctx.filter = 'url(#image)';
    ctx.fillRect(5, 5, 10, 10);
  }, 'url(#image)');
}, "Non-whitelisted SVG filters taint the canvas");
</script>
